fairness_curve/requirements.txt
beautifulsoup4==4.12.3
clip==0.2.0
matplotlib==3.8.0
numpy==2.0.1
openai==1.37.1
pandas==2.2.2
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.14.0
torch==2.4.0
tqdm==4.65.0
